requirements.txt
et_xmlfile==2.0.0
lxml==5.3.0
openpyxl==3.1.5
pillow==11.0.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
PyYAML==6.0.2
typing_extensions==4.12.2
XlsxWriter==3.2.0
